+++ src/include/graph/ArrayStochasticNode.h
@@ -12,8 +12,8 @@ class ArrayDist;
  */
 class ArrayStochasticNode : public StochasticNode {
     ArrayDist const * const _dist;
-    std::vector<std::vector<unsigned int> > _dims;
-    void sp(double *lower, double *upper, unsigned int length, 
+    std::vector<std::vector<unsigned long> > _dims;
+    void sp(double *lower, double *upper, unsigned long length, 
 	    unsigned int chain) const;
 public:
     /**
@@ -24,7 +24,7 @@ public:
     ArrayStochasticNode(ArrayDist const *dist, unsigned int nchain,
 			std::vector<Node const *> const &parameters,
 			Node const *lower, Node const *upper,
-			double const *data=0, unsigned int length=0);
+			double const *data=0, unsigned long length=0);
     double logDensity(unsigned int chain, PDFType type) const;
     void randomSample(RNG *rng, unsigned int chain);
     void truncatedSample(RNG *rng, unsigned int chain,
@@ -32,7 +32,7 @@ public:
     bool checkParentValues(unsigned int chain) const;
     //StochasticNode *clone(std::vector<Node const *> const &parents,
     //Node const *lower, Node const *upper) const;
-    unsigned int df() const;
+    unsigned long df() const;
     double KL(unsigned int chain1, unsigned int chain2, RNG *rng, unsigned int nrep) const;
 };
